Add files via upload
here is the fizz and here is buzz! at last, after many attempts at fizz and buzz, we have two new exotic features! the focus of this update was actually a tangent from the main project and part of a special pacific research project. with this, we hope to see how well squids respond to fizz.